chore: updates eleventy config
